hid-g19: fix crash on module or device removal
* reorder cleanup sequence
* better cleanup on probe errors

Still getting crashes about sysfs attributes when removing the hid-g19
and/or the hid-ginput module while hid-gfb is in use by userspace.

These can be fixed by proper ownership of attributes by the respective
drivers (fb attributes go with hid-gfb, input attributes with hid-ginput
etc), but this breaks the current API for userspace and will be fixed
later.